Apis/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart
pydantic
requests
passlib
